RegionOfInterest has offset.  RegionOfInterestOffset removed.
